Generalized"ExpandingButton" into "CompressibleUI"
So now Compressible UI can work on all UI elements, and can be added in
a simple manner.  Also de-singleton'd LeapInputModule because
BroadcastMessage exists.

Take note that Compressible UI elements, will only work correctly if the
same gameobject that holds it is the only one with RaycastTarget set to
true..