[FIX] Fix Script Compatibility with Unity 2017.4
The Unity Visualizer is built in Jenkins with Unity 2017.4; this PR fixes the Script-based incompatibilities with Unity 2017.4 (and presumably the versions of 2018.x in-between).

This PR does not attempt to address the Prefab and Scene Serialization issues with 2017.4 (which may still cause problems when building the Unity Visualizer).   To properly deal with these, we will probably need to update the version of Unity used in Jenkins (to 2019.2 or so).