Port Drawing namespace Sim-compatibility changes -> UnityPlugin
This commit moves the "Drawer" from the main Leap.Unity namespace to a new Leap.Unity.Drawing namespace. Drawer is only used in a handful of places in the plugin, and it's a very generic name, so importers of the Leap.Unity SDK shouldn't have that name conflict with one of their own classes.

This is a significant simplification for a number of files in the Simulator codebase, which is a consumer of the SDK's Leap.Unity namespace and, indeed, has its own Drawer class.

This commit also removes a handful of unused imports from the files it touches.