GEM-572 Fix workarounds in unitymodules for PolicyFlags.
Previously we were attempting to keep track of the state of flags
ourselves, but this requires making (invalid) assumptions about the
initial state of the flags.

As of libtrack MR 1002, the client will send over the correct delta
required for set/clear flags rather than sending flags as a whole. As
a result, maintaining the flags is no longer necessary.

It also erroneously wipes out old flags, for example setting the Image
flag would create a bitwise not of the Image flag, which could
accidentally set the tracking mode states back to default (such as
desktop).